requirements.txt
pytest>=7.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
psutil>=5.9.0
httpx>=0.24.0
applicationinsights>=0.11.10
sqlalchemy>=2.0.0
passlib[argon2]>=1.7.4
pyjwt>=2.8.0
email-validator>=2.1.0
